Revert "litres-60 feature(user-data) create abstractions"
This reverts commit 627f0fa086af8d8a13fb7562291311f276c16a59.
