+++ backend/src/Litres.Data/Configurations/Mapping/UserMapperProfile.cs
@@ -11,8 +11,5 @@ public class UserMapperProfile : Profile
         CreateMap<UserRegistrationDto, User>()
             .ForMember("UserName", opt => opt.MapFrom(dto => dto.Email))
             .ForMember("PasswordHash", opt => opt.MapFrom(dto => dto.Password));
-
-        CreateMap<UserSettingsDto, User>();
-        CreateMap<User, UserSettingsDto>();
     }
 }
