--- backend/src/Litres.Data/Dto/Requests/UserSettingsDto.cs
@@ -1,7 +1,0 @@
-namespace Litres.Data.Dto.Requests;
-
-public class UserSettingsDto
-{
-    public string Name { get; set; }
-    public string AvatarUrl { get; set; }
-}
