+++ backend/src/Litres.Main/Extensions/WebApplicationExtension.cs
@@ -17,8 +17,7 @@ public static class WebApplicationExtension
         #region Services
         builder.Services.AddScoped<IRequestService, RequestService>();
         builder.Services.AddScoped<IBookService, BookService>();
-        builder.Services.AddScoped<IRegistrationService, RegistrationService>();
-        builder.Services.AddScoped<IUserService, UserService>();
+        builder.Services.AddScoped<IRegistrationService, UserService>();
         #endregion
         
         #region Repositories
